requirements.txt
django
pillow
unidecode
coverage
django-baton 2.2.1
django-admin-numeric-filter
django-imagekit
model_mommy
djangorestframework
markdown
django-filter
restframework-simplejwt
djoser
social-auth-app-django
